'req.param' was deprecated. Replace with 'req.query'
Summary:

Express.js deprecated `req.param` in favor of `req.query`.

Changed semantics slightly.
`req.param` used to be a function that would fetch the supplied param.
`req.query` is a function that returns a JSON object of all params and
values.

Test Plan:

This would only happen on subscription since that's where this was used.

Patched the change to a test app instance on Heroku.
Deployed.
Added a new subscription.
Did not see deprecation logs anymore.